+++ react/components/OrganizationUsersTable.tsx
@@ -123,10 +123,6 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
     'impersonate-users-organization'
   )
 
-  const canImpersonateCost = permissions.includes(
-    'impersonate-users-costcenter'
-  )
-
   const { data, loading, refetch } = useQuery(GET_USERS, {
     variables: {
       ...initialState,
@@ -144,6 +140,10 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
     ssr: false,
     skip: isAdmin,
   })
+
+  const canImpersonateCost = (rowData: B2BUserSimple) =>
+    permissions.includes('impersonate-users-costcenter') &&
+    rowData.costId === costCenterData?.getCostCenterByIdStorefront?.id
 
   const [addUser] = useMutation(ADD_USER)
   const [updateUser] = useMutation(UPDATE_USER)
@@ -329,29 +329,9 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
 
   const handleImpersonation = (rowData: B2BUserSimple) => {
     if (
-      (!canImpersonateAll && !canImpersonateOrg && !canImpersonateCost) ||
-      rowData.role.slug.includes('sales')
-    ) {
-      showToast(formatMessage(storeMessages.toastImpersonateForbidden))
-
-      return
-    }
-
-    if (
-      !canImpersonateAll &&
-      canImpersonateOrg &&
-      !rowData.role.slug.includes('customer')
-    ) {
-      showToast(formatMessage(storeMessages.toastImpersonateForbidden))
-
-      return
-    }
-
-    if (
       !canImpersonateAll &&
       !canImpersonateOrg &&
-      canImpersonateCost &&
-      rowData.costId !== costCenterData?.getCostCenterByIdStorefront?.id
+      !canImpersonateCost(rowData)
     ) {
       showToast(formatMessage(storeMessages.toastImpersonateForbidden))