+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.27.3] - 2023-10-10
+
+### Fixed
+
+- Improve quality gate on Sonar
+
+## [1.27.2] - 2023-10-06
+
+### Fixed
+
+- Fix to impersonate user with the role `sales`
+
 ## [1.27.1] - 2023-08-17
 
 ### Fix